wamp client remade by unix socket
