+++ example/wamp_client/b.py
@@ -1,14 +1,22 @@
 from wampify import Wampify
-from wampify.wamp_client import WAMPClient
+from wampify._wamp_client import WAMPClient
 from wampify.signal_manager import wamps_signals
+import asyncio
 
 
 ANOTHER_ROUTER_URL = 'ws://127.0.0.1:8080/private'
-client = WAMPClient(router_url=ANOTHER_ROUTER_URL)
+client = WAMPClient(router_url=ANOTHER_ROUTER_URL, realm='example')
 
 @wamps_signals.on
 async def joined(session, details):
-    await client.join(realm='example')
+    await client.join()
+    async def f(i):
+        return i ** 2 == await client.call('com.example.pow', i)
+    print('wamp client session joined')
+    print(await asyncio.gather(
+        *[f(i) for i in range(3)],
+        return_exceptions=True
+    ))
 
 @wamps_signals.on
 async def leaved(session, details):
@@ -29,8 +37,6 @@ wampify = Wampify(
 @wampify.subscribe
 async def hello(name: str = 'Anonymous'):
     print(f'{name} you are welcome!')
-    answer = await client.call('com.example.pow', 10)
-    print(f'Another session answered: {answer}')
 
 
 if __name__ == '__main__':
